requirements.txt
discord.py
python-dotenv
cassandra-driver
dateparser